requirements.txt
absl-py==1.4.0
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
clean-fid==0.1.35
contextlib2==21.6.0
filelock==3.9.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
lmdb==1.4.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
ml-collections==0.1.1
mpi4py
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.68
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.17
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
